cpuidle: Wrap cpu_online_mask usage with preempt_disable
Make sure preemption is disabled before we read the cpu_online_mask
to ensure that no CPUs are hotplugged out after. If we do not have
preemption disabled when we read the online mask, we run the risk of
sending an IPI to an offline CPU resulting in a BUG().

Change-Id: I349cfd0ff3d269f3d544b458eaad27b627470c09