kernel: power: qos: remove check for core isolation while cluster LPMs
Since all cores in a cluster are in isolation, PMQoS latency constraint
set by clock driver to switch PLL is ignored. So, Cluster enter to L2PC
and SPM is trying to disable the PLL and at same time clock driver
trying to switch the PLL from other cluster which leads to the
synchronization issues.

Fix is although all cores are in isolation, honor PMQoS request
for cluster LPMs.

Change-Id: I4296e16ef4e9046d1fbe3b7378e9f61a2f11c74d