msm: kgsl: Use kthread instead of workqueue for event work
Currently a workqueue is being used to process the event work. In certain
scenarios like when most of CPU cores are busy, there can be a significant
delay between the actual timestamp retire event and when the work is
processed by the events workqueue as workqueues cannot have RT priority.
Hence use kthread instead of workqueue for event work.

Change-Id: Ib1ec7fa1ec3a133d03104c9a029dcc4c06180609