sched: Implement tapered dvfs_headroom
The linear one is very aggressive at the top frequencies. 25% headroom
means that we saturate the cluster very quickly. Given how expensive the
new clusters are at the top frequencies, it'd be smarter to taper the
headroom as we approach the max capacity of the cluster.

Implement that by simply taking the headroom of the different between
curret util and capacity of the CPU. The lower frequencies will get
a big headroom, but as we approach max capacity, we'll get almost no
headroom.

Only CFS + RT utilization use the tapered version. irq is usually small
and this can lead to potentially undesired boost almost when there are
no irqs running. DL shouldn't need any headroom in general. So keeping
it using the default non tapered version as well.

Ported from: LineageOS/android_kernel_google_gs201@752c5f9

Signed-off-by: Qais Yousef <qyousef@google.com>
Signed-off-by: Dark-Matter7232 <me@const.eu.org>